requirements.txt
streamlit
pandas
PyGithub
feedparser
janome
requests
